requirements/test.txt
# Test dependencies go here.
-r base.txt

coverage==4.3.4
django-coverage-plugin==1.5.0
django-test-plus==1.0.17
factory-boy==2.8.1
selenium==3.4.2

flake8==3.3.0
flake8-docstrings==1.1.0
flake8-import-order==0.12

# pytest
pytest-django==3.1.2
pytest-sugar==0.8.0
